Create Expandable  Pytorch Sine-x Model
This is an expandable pytorch-based version of his tensorflow model.py.  This pytorch version provides Hyperparameter Dictionary control of the size and depth of hidden layers and swapping of Activation Functions, for evaluating model optimization.   Pytorch also implements a checkpoint.pth file export feature (not implemented in this script) to obtain the set of weights that are shown to provide optimal performance.